Room decompilation mostly complete. Several rooms/scenes tested and matching.

+++ ZAP2/BitConverter.h
@@ -18,8 +18,8 @@ public:
 		return (data[offset + 0] << 24) + (data[offset + 1] << 16) + (data[offset + 2] << 8) + data[offset + 3];
 	}
 
-	static int64_t ToInt64BE(std::vector<uint8_t> data, int offset)
+	static uint64_t ToInt64BE(std::vector<uint8_t> data, int offset)
 	{
-		return (data[offset + 0] << 56) + (data[offset + 1] << 48) + (data[offset + 2] << 40) + (data[offset + 3] << 32) + (data[offset + 4] << 24) + (data[offset + 5] << 16) + (data[offset + 6] << 8) + (data[offset + 7]);
+		return ((uint64_t)data[offset + 0] << 56) + ((uint64_t)data[offset + 1] << 48) + ((uint64_t)data[offset + 2] << 40) + ((uint64_t)data[offset + 3] << 32) + ((uint64_t)data[offset + 4] << 24) + ((uint64_t)data[offset + 5] << 16) + ((uint64_t)data[offset + 6] << 8) + ((uint64_t)data[offset + 7]);
 	}
 };
